+++ src/js/util/position.js
@@ -39,7 +39,7 @@ function attachTo(element, target, options) {
     const position = offset(element);
 
     for (const [i, [prop, dir, start, end]] of Object.entries(dirs)) {
-        const targetOffset = offsetViewport(target[i]);
+        const targetOffset = offset(target[i]);
         position[start] = position[dir] =
             targetOffset[start] +
             moveBy(attach.target[i], end, targetOffset[prop]) -